hal: Add ultrasound proximity support
The sensors.ultrasoundproximity.so sensors module notifies audio
HAL to enable the ultrasound stream through the audio_us_socket_0
socket. The ultrasound extension listens for messages on this
socket and sets the ultrasound-proximity parameter accordingly.
Then, it starts the ultrasound stream and registers a callback to
handle the proximity events. These events are passed to the
libultrasound_notify lib which communicates with the sensors module
and reports the sensor readings in the end.

Change-Id: I9e005d6ff8f414a675e1ac89ab4c13c61a78205b